refactor - move SSRNET specific model definitions to its own module
